pyproject.toml
[tool.poetry]
name = "channel-order"
version = "0.1.0"
description = ""

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.6.0"
torchvision = "^0.7.0"
pytorch-lightning = "^1.0.2"
opencv-python = "^4.4.0"
numpy = "^1.19.2"
ttach = "^0.0.3"
tensorboard = "^2.3.0"
albumentations = "^0.4.6"
efficientnet-pytorch = "^0.7.0"
pyserde = {extras = ["yaml"], version = "^0.1.5"}
test-tube = "^0.7.5"

[tool.poetry.dev-dependencies]
mypy = "^0.790"
flake8 = "^3.8.4"
black = "^20.8b1"
pytest = "^6.1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

